LRCI-6021 Make DownstreamBuild into an interface
