+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/DefaultDownstreamBuild.java
@@ -1,0 +1,17 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jenkins.results.parser;
+
+/**
+ * @author Michael Hashimoto
+ */
+public class DefaultDownstreamBuild extends BaseDownstreamBuild {
+
+	protected DefaultDownstreamBuild(String url, TopLevelBuild topLevelBuild) {
+		super(url, topLevelBuild);
+	}
+
+}
