+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PoshiDownstreamBuild.java
@@ -17,7 +17,7 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public class PoshiDownstreamBuild extends DownstreamBuild {
+public class PoshiDownstreamBuild extends BaseDownstreamBuild {
 
 	@Override
 	public List<TestResult> getTestResults() {
